utils: Directly use syscall() to close open FDs in closefrom()
Using malloc() after fork() is potentially unsafe, so we should avoid it
if possible.

WIP: we might want to remove the unsafe version (or at least fix some
aspects of it, e.g. reduce allocations by directly calling opendir()
and readdir() and make sure the FD used by opendir(), if any, is not closed).

 #990.